requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=py311haa95532_0
asttokens=3.0.0=pypi_0
attrs=25.3.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
blas=1.0=mkl
bleach=6.2.0=pypi_0
bottleneck=1.4.2=py311h57dcf0c_0
brotli-python=1.0.9=py311h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=py311haa95532_0
contourpy=1.3.1=py311h214f63a_0
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=py311haa95532_0
docopt=0.6.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.13.1=py311haa95532_0
fonttools=4.55.3=py311h827c3e9_0
freetype=2.13.3=h0620614_0
fsspec=2024.12.0=py311haa95532_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py311h827c3e9_0
grpcio=1.71.0=py311hbb1fe6b_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipython=8.12.3=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=py311haa95532_0
joblib=1.4.2=py311haa95532_0
jpeg=9e=h827c3e9_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.8=py311h5da7b33_0
kornia=0.8.0=pypi_0
kornia-rs=0.1.8=pypi_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libabseil=20250127.0=cxx17_h52369b4_0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libgrpc=1.71.0=hf4237ab_0
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h70ee33d_0
libprotobuf=5.29.3=hac4c8cb_0
libre2-11=2024.07.02=h5da7b33_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h3d04722_1
lightning-utilities=0.11.9=py311haa95532_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.4.1=py311haa95532_0
markupsafe=3.0.2=py311h827c3e9_0
matplotlib=3.10.0=py311haa95532_0
matplotlib-base=3.10.0=py311he19b0ae_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.3=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h827c3e9_2
mkl_fft=1.3.11=py311h827c3e9_0
mkl_random=1.2.8=py311hea22821_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py311haa95532_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
networkx=3.4.2=py311haa95532_0
numexpr=2.10.1=py311h4cd664f_0
numpy=1.26.4=py311hdab7c0b_0
numpy-base=1.26.4=py311hd01c5d8_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.16=h3f729d1_0
opentelemetry-api=1.30.0=py311haa95532_0
packaging=24.2=py311haa95532_0
pandas=2.2.3=py311h5da7b33_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=11.1.0=py311h096bfcc_0
pip=25.0=py311haa95532_0
pipreqs=0.5.0=pypi_0
platformdirs=4.3.7=pypi_0
prompt-toolkit=3.0.51=pypi_0
protobuf=5.29.3=py311h5da7b33_0
pure-eval=0.2.3=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.0=py311haa95532_0
pyqt=6.7.1=py311h5da7b33_0
pyqt6-sip=13.9.1=py311h827c3e9_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0post0=py311haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.3.1=cpu_py311h0ce1571_0
pytz=2024.1=py311haa95532_0
pywin32=310=pypi_0
pyyaml=6.0.2=py311h827c3e9_0
pyzmq=26.4.0=pypi_0
qtbase=6.7.3=h0804d20_0
qtdeclarative=6.7.3=h5da7b33_0
qtsvg=6.7.3=hf2fb9eb_0
qttools=6.7.3=h0de5f00_0
qtwebchannel=6.7.3=h5da7b33_0
qtwebsockets=6.7.3=h5da7b33_0
re2=2024.07.02=h214f63a_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.24.0=pypi_0
scikit-learn=1.6.1=py311h585ebfc_0
scipy=1.15.2=py311h9d85e7c_1
seaborn=0.13.2=py311haa95532_2
setuptools=75.8.0=py311haa95532_0
sip=6.10.0=py311h5da7b33_0
six=1.17.0=py311haa95532_0
sleef=3.5.1=h8cc25b3_2
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
sympy=1.13.3=py311haa95532_1
tbb=2021.8.0=h59b6b97_0
tensorboard=2.19.0=py311haa95532_0
tensorboard-data-server=0.7.0=py311haa95532_1
threadpoolctl=3.5.0=py311h746a85d_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h0416ee5_0
torchmetrics=1.4.0.post0=py311haa95532_0
torchvision=0.18.1=cpu_py311h9788fc5_0
tornado=6.4.2=py311h827c3e9_0
tqdm=4.67.1=py311h746a85d_0
traitlets=5.14.3=pypi_0
typing_extensions=4.12.2=py311haa95532_0
tzdata=2025a=h04d1e81_0
unicodedata2=15.1.0=py311h827c3e9_1
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.1.3=py311haa95532_0
wheel=0.45.1=py311haa95532_0
wrapt=1.17.0=py311h827c3e9_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
yarg=0.1.9=pypi_0
zipp=3.21.0=py311haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
